requirements.txt
# LangChain ecosystem
langchain>=0.3.25
langchain-core>=0.3.63
langchain-community>=0.3.24
langchain-text-splitters>=0.3.8
langgraph>=1.0.8

# Vector store
langchain-chroma>=0.2.4
chromadb>=1.0.9

# Graph store
langchain-neo4j>=0.4.0
neo4j>=5.28.1

# Local LLM
langchain-ollama>=0.3.3
ollama>=0.6.1

# Utilities
python-dotenv>=1.2.1
pydantic>=2.10.0
rich>=13.9.0
typing-extensions>=4.12.0

# Code parsing
astroid>=3.3.0
tree-sitter>=0.24.0
tree-sitter-python>=0.24.0
